phase-05-cloud-deploy/services/backend-api/requirements.txt
fastapi>=0.109.0
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic-settings>=2.1.0
requests>=2.31.0
python-jose[cryptography]>=3.3.0
pytest>=7.4.0
httpx>=0.25.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0
openai>=1.0.0
dapr>=1.12.0
aiokafka>=0.10.0
resend>=0.7.0
asyncpg>=0.29.0
